@@ -47,7 +47,7 @@
     "casual": "^1.6.2",
     "core-js": "^3.20.2",
     "es6-promise": "^4.2.8",
-    "fuse.js": "^6.6.2",
+    "fuse.js": "^7.0.0",
     "history": "^5.3.0",
     "isomorphic-fetch": "^3.0.0",
     "jquery": "^3.5.1",
@@ -61,12 +61,12 @@
     "posthog-js": "^1.225.0",
     "promise-polyfill": "^8.3.0",
     "ramda": "^0.30.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.1",
+    "react": "^18.2.14",
+    "react-bootstrap": "^2.0.0",
     "react-content-loader": "^7.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.14",
     "react-dotdotdot": "1.3.1",
-    "react-infinite-scroller": "^1.2.4",
+    "react-infinite-scroller": "^1.2.6",
     "sass": "1.77.8",
     "video.js": "^7.11.8",
     "videojs-transcript-ac": "^0.8.8",
@@ -74,12 +74,12 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/preset-typescript": "^7.25.7",
     "@mitodl/ocw-to-hugo": "1.36.0",
     "@playwright/test": "^1.45.0",
     "@swc/core": "^1.3.14",
     "@swc/helpers": "^0.4.12",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@types/enzyme": "^3.10.11",
+    "@testing-library/react": "15",
     "@types/inquirer": "^8.2.5",
     "@types/jest": "^27.4.0",
     "@types/jquery": "^3.5.11",
@@ -87,14 +87,13 @@
     "@types/lodash.uppercase": "^4.3.6",
     "@types/promise-polyfill": "^6.0.4",
     "@types/ramda": "^0.27.62",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-infinite-scroller": "^1.2.2",
+    "@types/react": "^18.2.14",
+    "@types/react-dom": "^18.2.14",
+    "@types/react-infinite-scroller": "^1.2.5",
     "@types/serve-handler": "^6.1.1",
     "@types/sinon": "^17.0.0",
     "@types/video.js": "^7.3.29",
     "@typescript-eslint/eslint-plugin": "^5.40.1",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.4.1",
     "ansi-colors": "^4.1.3",
     "clean-webpack-plugin": "^4.0.0",
     "cli-table3": "^0.6.3",
